refactored tests into folders
