+++ Backend/src/test/java/com/Backend/Backend/EntityTests/FlavorTest.java
@@ -1,4 +1,4 @@
-package com.Backend.Backend;
+package com.Backend.Backend.EntityTests;
 
 import com.Backend.Backend.entity.Flavor;
 import org.junit.jupiter.api.Test;
