+++ Backend/src/test/java/com/Backend/Backend/SecurityTests/JwtAuthFilterTest.java
@@ -1,4 +1,4 @@
-package com.Backend.Backend;
+package com.Backend.Backend.SecurityTests;
 
 import com.Backend.Backend.security.JwtAuthFilter;
 import com.Backend.Backend.security.JwtUtils;
